+++ src/main/kotlin/io/github/freya022/botcommands/api/commands/application/ApplicationCommandInfo.kt
@@ -1,7 +1,7 @@
 package io.github.freya022.botcommands.api.commands.application
 
 import io.github.freya022.botcommands.api.commands.CommandInfo
-import io.github.freya022.botcommands.api.commands.Executable
+import io.github.freya022.botcommands.api.core.Executable
 
 interface ApplicationCommandInfo : CommandInfo, Executable {
     val topLevelInstance: TopLevelApplicationCommandInfo
