+++ src/main/kotlin/io/github/freya022/botcommands/api/commands/text/TextCommandVariation.kt
@@ -1,7 +1,7 @@
 package io.github.freya022.botcommands.api.commands.text
 
-import io.github.freya022.botcommands.api.commands.Executable
 import io.github.freya022.botcommands.api.commands.builder.IDeclarationSiteHolder
+import io.github.freya022.botcommands.api.core.Executable
 
 interface TextCommandVariation : Executable, IDeclarationSiteHolder {
     override val parameters: List<TextCommandParameter>
